backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "individual-teacher-backend"
version = "1.0.0"
authors = [
  { name="Individual Teacher Team" }
]
description = "AI Individual Teacher Backend API"
requires-python = ">=3.10"
dependencies = [
  "Flask==3.0.3",
  "Flask-Cors==4.0.0",
  "python-dotenv==1.0.1",
  "pymongo[srv]==4.7.0",
  "requests==2.31.0",
  "numpy>=1.26.4,<2.0",
  "scikit-learn==1.4.2",
  "langchain==0.1.17",
  "langchain-groq==0.1.3",
  "pytest==8.1.1",
  "pytest-asyncio==0.23.6",
  "huggingface-hub>=0.22.2",
  "sentence-transformers>=2.7.0",
  "pydantic==2.9.2",
  "python-multipart==0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1",
  "pytest-asyncio>=0.23.6"
]

[tool.setuptools]
packages = ["api", "core", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]